@@ -5,7 +5,7 @@
     <parent>
         <groupId>org.springframework.boot</groupId>
         <artifactId>spring-boot-starter-parent</artifactId>
-        <version>2.1.6.RELEASE</version>
+        <version>3.1.0</version>
         <relativePath/>
     </parent>
     <groupId>com.madadipouya.redis.springdata.example</groupId>
@@ -14,8 +14,8 @@
     <name>redis-springdata</name>
     <description>Spring Data Redis Example</description>
     <properties>
-        <java.version>1.8</java.version>
-        <kotlin.version>1.3.40</kotlin.version>
+        <java.version>17</java.version>
+        <kotlin.version>1.8.21</kotlin.version>
     </properties>
     <dependencies>
         <dependency>
@@ -24,13 +24,16 @@
         </dependency>
         <dependency>
             <groupId>org.springframework.boot</groupId>
+            <artifactId>spring-boot-starter-validation</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.springframework.boot</groupId>
             <artifactId>spring-boot-starter-data-redis</artifactId>
         </dependency>
         <dependency>
             <groupId>redis.clients</groupId>
             <artifactId>jedis</artifactId>
-            <version>2.9.0</version>
-            <type>jar</type>
+            <version>4.4.1</version>
         </dependency>
         <dependency>
             <groupId>org.jetbrains.kotlin</groupId>
@@ -43,7 +46,7 @@
         <dependency>
             <groupId>com.fasterxml.jackson.module</groupId>
             <artifactId>jackson-module-kotlin</artifactId>
-            <version>2.9.9</version>
+            <version>2.15.2</version>
         </dependency>
         <dependency>
             <groupId>io.springfox</groupId>
